+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - CASMCMS-8795 - Updated for remote builds.
+- CASMCMS-8818 - ssh key injection into jobs.
 
 ## [1.6.0] - 2023-09-15
 ### Changed